\name{print.predict_bayesCureModel}
\alias{print.predict_bayesCureModel}
\title{
Print method for the \code{predict} object
}
\description{
This function prints a summary of objects returned by the \code{predict.cure_rate_MC3} method.
}
\usage{
\method{print}{predict_bayesCureModel}(x, ...)
}
\arguments{
  \item{x}{
An object of class \code{predict_bayesCureModel}, which is returned by the \code{predict.cure_rate_MC3} method.
}
  \item{...}{
        ignored.
}
}
\value{No return value, called for side effects.}
\details{
The function prints some basic information for the \code{predict} method of a \code{bayesCureModel} object. 
}
\author{
Panagiotis Papastamoulis
}

