\name{print.summary_bayesCureModel}
\alias{print.summary_bayesCureModel}
\title{
Print method for the summary
}
\description{
This function prints a summary of objects returned by the \code{summary.cure_rate_MC3} method.
}
\usage{
\method{print}{summary_bayesCureModel}(x, digits = 2, ...)
}
\arguments{
  \item{x}{
An object of class \code{summary_bayesCureModel}, which is returned by the \code{summary.cure_rate_MC3} method.
}
  \item{digits}{
Number of digits to print.
}
  \item{...}{
        ignored.
}
}
\value{No return value, called for side effects.}
\details{
The function prints some basic information for the summary of a \code{bayesCureModel} object. 
}
\author{
Panagiotis Papastamoulis
}

